Add a new, simpler interface for accessing height info
FEATURE: The new `elementAtHeight`, `lineBlockAtHeight`, and `lineBlockAt` methods
provide a simpler and more efficient replacement for the (now deprecated) `blockAtHeight`,
`visualLineAtHeight`, and `visualLineAt` methods.

FEATURE: The editor view now exports a `documentTop` getter that gives you the vertical
position of the top of the document. All height info is queried and reported relative to
this top.

FEATURE: The editor view's new `viewportLineBlocks` property provides an array of
in-viewport line blocks, and replaces the (now deprecated) `viewportLines` method.