Add @codemirror/state to peerDependencies
This tells package manager that this package uses @codemirror/state
as part of its public API and should try to make sure the package
is deduped to avoid issues with duplicate packages.

The peer dependency here duplicates a regular dependency entry. This
is intentional to preserve compatibility with old npm versions that
did not automatically install peer dependencies.