Allow multiple transactions to be dispatched at the same time
FEATURE: `EditorView.dispatch` now also accepts an array of transactions to be applied
together in a single view update.

FEATURE: The new `dispatchTransactions` option to `new EditorView` now replaces the old
(deprecated but still supported) `dispatch` option in a way that allows multiple
transactions to be applied in one update.